enables installer to also set up a browser in kiosk mode

+++ bin/updateopenrowingmonitor.sh
@@ -11,7 +11,7 @@ set -u
 set -e
 
 print() {
-  printf "%s\n" "$@"
+  echo "$@"
 }
 
 cancel() {
@@ -19,8 +19,38 @@ cancel() {
   exit 1
 }
 
+ask() {
+  local prompt default reply
+
+  if [[ ${2:-} = 'Y' ]]; then
+    prompt='Y/n'
+    default='Y'
+  elif [[ ${2:-} = 'N' ]]; then
+    prompt='y/N'
+    default='N'
+  else
+    prompt='y/n'
+    default=''
+  fi
+
+  while true; do
+    echo -n "$1 [$prompt] "
+    read -r reply </dev/tty
+
+    if [[ -z $reply ]]; then
+      reply=$default
+    fi
+
+    case "$reply" in
+      Y*|y*) return 0 ;;
+      N*|n*) return 1 ;;
+    esac
+  done
+}
+
 CURRENT_DIR=$(pwd)
-INSTALL_DIR="/opt/openrowingmonitor"
+SCRIPT_DIR="$( cd "$(dirname "${BASH_SOURCE[0]}")" &> /dev/null && pwd )"
+INSTALL_DIR="$(dirname "$SCRIPT_DIR")"
 GIT_REMOTE="https://github.com/laberning/openrowingmonitor.git"
 
 print "Update script for Open Rowing Monitor"
@@ -35,9 +65,7 @@ REMOTE_VERSION=$(git ls-remote $GIT_REMOTE HEAD | awk '{print $1;}')
 if [ "$LOCAL_VERSION" = "$REMOTE_VERSION" ]; then
     print "You are using the latest version of Open Rowing Monitor."
 else
-    print "A new version of Open Rowing Monitor is available. Do you want to update?"
-    print
-    read -p "Press RETURN to continue or CTRL + C to abort"
+  if ask "A new version of Open Rowing Monitor is available. Do you want to update?" Y; then
     print "Stopping Open Rowing Monitor..."
     sudo systemctl stop openrowingmonitor
 
@@ -56,6 +84,7 @@ else
     print
     print "Update complete, Open Rowing Monitor now has the following exciting new features:"
     git log --reverse --pretty=format:"- %s" $LOCAL_VERSION..HEAD
+  fi
 fi
 
 cd $CURRENT_DIR
